accounting/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-mysql==4.11.0
django-cors-headers==4.2.0
drf-yasg==1.21.7
mysqlclient==2.2.0
python-dotenv==1.0.0
pika==1.1.0
